[Tool] Support compilation on Mac 9: be code 3 (#63853)

+++ be/src/http/action/runtime_filter_cache_action.cpp
@@ -78,8 +78,8 @@ void RuntimeFilterCacheAction::_handle_stat(HttpRequest* req) {
     const std::string& enable_trace = _exec_env->runtime_filter_cache()->enable_trace() ? "true" : "false";
     _handle(req, [=](rapidjson::Document& root) {
         auto& allocator = root.GetAllocator();
-        root.AddMember("cache_times", rapidjson::Value(cache_times), allocator);
-        root.AddMember("use_times", rapidjson::Value(use_times), allocator);
+        root.AddMember("cache_times", rapidjson::Value(static_cast<uint64_t>(cache_times)), allocator);
+        root.AddMember("use_times", rapidjson::Value(static_cast<uint64_t>(use_times)), allocator);
         root.AddMember("enable_trace", rapidjson::Value(enable_trace.c_str(), enable_trace.size()), allocator);
     });
 }
